rateprofessor/requirements.txt
Python 3.10.0
Django==5.1.6
djangorestframework==3.15.2
requests==2.32.3